Split target implementations into separate files
